Combine duplicated articles
The introductory articles for properties and indexers were largely duplicated, combine them.
