+++ .openpublishing.redirection.csharp.json
@@ -272,6 +272,10 @@
             "redirect_url": "/dotnet/csharp/language-reference/operators/lambda-expressions"
         },
         {
+            "source_path_from_root": "/docs/csharp/indexers.md",
+            "redirect_url": "/dotnet/csharp/programming-guide/indexers"
+        },
+        {
             "source_path_from_root": "/docs/csharp/interactive-with-bash.md",
             "redirect_url": "/dotnet/csharp/index"
         },
@@ -3815,6 +3819,10 @@
         {
             "source_path_from_root": "/docs/csharp/programming-guide/xmldoc/xml-documentation-comments.md",
             "redirect_url": "/dotnet/csharp/programming-guide/xmldoc"
+        },
+        {
+            "source_path_from_root": "/docs/csharp/properties.md",
+            "redirect_url": "/dotnet/csharp/programming-guide/classes-and-structs/properties"
         },
         {
             "source_path_from_root": "/docs/csharp/quick-starts/arrays-and-collections.md",
