+++ docs/csharp/toc.yml
@@ -287,10 +287,6 @@ items:
     href: nullable-migration-strategies.md
   - name: Methods
     href: methods.md
-  - name: Properties
-    href: properties.md
-  - name: Indexers
-    href: indexers.md
   - name: Iterators
     href: iterators.md
   - name: Delegates & events
